wip: First refactoring pass
I'm not that happy with the Helpers approach
for now, however there's probably not any
alternative that is significantly more
elegant than what we currently have.
The things I'm most unsure about is _where_
to place the unwrapping calls. A lot of
methods that are protected for example
seem like they're intended for internal
use primarily, however they're not private.

What's left to do for now as a first step
is to add the respective tests so we can
evaluate this implementation for obvious
flaws. Afterwards we can still think about
whether we can make it more elegant.